Clean: Clean indent, ts interface
