+++ packages/plugin-xrpl/src/services/getTransactionService.ts
@@ -2,22 +2,7 @@ import { elizaLogger } from "@elizaos/core";
 import { dropsToXrp } from "xrpl";
 import { xrplClient } from "./lib/xrplClient";
 
-interface TransactionParams {
-	address: string;
-	transactionCount?: number;
-	startDate?: string;
-	endDate?: string;
-}
 
-interface Transaction {
-	type: string;
-	hash: string;
-	date: string;
-	amount: string;
-	from: string;
-	to: string;
-	status: string;
-}
 
 export async function getTransactionService({ 
 	address,
