+++ packages/plugin-xrpl/src/utils/formatDate.ts
@@ -1,0 +1,10 @@
+export function formatDate(dateString: string): string {
+	const date = new Date(dateString);
+	return date.toLocaleString('en-US', {
+		year: 'numeric',
+		month: 'short',
+		day: 'numeric',
+		hour: '2-digit',
+		minute: '2-digit'
+	});
+}
